Extract shared metadata-change hook from Topbar
The draft-status helper needs the same save-to-server/idb/store path the
topbar uses; useMetadataChange is that path as a hook.
