+++ app/src/app/components/Topbar/Topbar.tsx
@@ -9,8 +9,7 @@ import {defaultPanels} from '@components/sidebar/DataPanelUtils';
 import {PasswordPromptModal} from '../Toolbar/PasswordPromptModal';
 import {UploaderModal} from '../Toolbar/UploaderModal';
 import {MapHeader} from './MapHeader';
-import {saveMapDocumentMetadata} from '@/app/utils/api/apiHandlers/saveMapDocumentMetadata';
-import {idb} from '@/app/utils/idb/idb';
+import {useMetadataChange} from '@/app/hooks/useMetadataChange';
 import {useMapControlsStore} from '@/app/store/mapControlsStore';
 import {MAP_MODES} from '@constants/map/mode';
 import {useIsDesktop} from '@/app/hooks/useIsDesktop';
@@ -25,26 +24,7 @@ export const Topbar: React.FC = () => {
   const [modalOpen, setModalOpen] = React.useState<'upload' | null>(null);
   const mapDocument = useMapStore(state => state.mapDocument);
   const isEval = useMapControlsStore(state => state.isEval);
-  const setNotification = useMapStore(state => state.setNotification);
-  const updateMetadata = useMapStore(state => state.updateMetadata);
-
-  const handleMetadataChange = async (updates: Partial<DocumentMetadata>) => {
-    if (!mapDocument?.document_id) return;
-    const response = await saveMapDocumentMetadata({
-      document_id: mapDocument?.document_id,
-      metadata: updates,
-    });
-    if (response.ok) {
-      idb.updateIdbMetadata(mapDocument?.document_id, updates);
-      updateMetadata(updates);
-    } else {
-      setNotification({
-        message: 'Failed to save metadata',
-        importance: 2,
-        type: 'error',
-      });
-    }
-  };
+  const handleMetadataChange = useMetadataChange();
 
   return (
     <>
